fix(linux): io-uring 架构白名单 + 32-bit libc 字段位宽
- io-uring 0.7 预编译 sys.rs 仅覆盖 x86_64/aarch64/riscv64/loongarch64/powerpc64
  原代码仅判 target_os=linux 让下游 i686/armv7 交叉编译触发 compile_error
  Cargo.toml/mod.rs/linux_io_uring.rs/lib.rs 四处同步加架构白名单 cfg
- 屏障移除后暴露的预先存在 32-bit libc 类型不匹配（一并修复）:
  - fs_detect.rs:128 statfs.f_type 在 i686 是 i32 → as i64
  - lib.rs:1574-1575 st_Xtime_nsec 在 i686 是 i32 → as i64
  - lib.rs:1582 st_ino 在 i686 是 u32 → as u64
- 验证：cargo check 通过 4 target
  (win-x86_64, linux-x86_64, linux-i686, linux-armv7)

+++ src/core/fs_detect.rs
@@ -125,7 +125,7 @@ fn statfs_strategy(path: &Path) -> Option<IoStrategy> {
     if rc < 0 {
         return None;
     }
-    let preliminary = strategy_from_ftype(sfs.f_type);
+    let preliminary = strategy_from_ftype(sfs.f_type as i64);
     if preliminary != IoStrategy::NetworkAsync {
         return Some(preliminary);
     }
